Revert "Fix cron delete task (#68)"
This reverts commit 168a5525b0bc55359764c8853654668ca1d27cc8.

+++ server/cron.js
@@ -7,11 +7,16 @@ module.exports = function(){
     }, 900000); // Every 15 minutes
   }
 
-  // Remove all entries marked as deleted over 31 days ago
-  var Sequelize = require('sequelize'),
-    db = require('./db')(Sequelize),
-    entryModel = require('./models/entry-model')(db, Sequelize),
-    entryService = new (require('./services/entry-service'))(entryModel, db);
+  // On the 31st of the 7 months/year that have 31 days,
+  // remove all entries marked as deleted in the database.
+  // I wait 31+ days because I need to outlast a maximum
+  // session lifetime which is currently 30 days.
+  if(new Date().getDate() === 31){
+    var Sequelize = require('sequelize'),
+      db = require('./db')(Sequelize),
+      entryModel = require('./models/entry-model')(db, Sequelize),
+      entryService = new (require('./services/entry-service'))(entryModel, db);
 
-  entryService.removeAllDeletedEntries();
+    entryService.removeAllDeletedEntries();
+  }
 }
